Add user API functionality and paths
Add user API functionality:

- UserList (GET /api/users):
To access this endpoint, a user must be authenticated (AnonymousUser requests are denied).
When this endpoint is accessed by an administrative user, it shows each registered user account using the UserSerializer.
When this endpoint is accessed by a normal user, it shows the requesting user's account details using the UserSerializer.

- UserCreateOrLogin (POST /api/users/create/)
This endpoint can only be accessed by AnonymousUsers. Authenticated users are rejected (HTTP 400 Bad request).
When this endpoint is accessed by an AnonymousUser and sttings.ALLOW_REGISTER is set to True, it is tried to create a user with given username and password through the request's body.
If a user with given username already exists, it is tried to perform a login to this user account.
If either registering or logging in is successful, the endpoint returns the user's username for validation and the authentication token.

- UserDetail (/api/users/<pk>/)
This endpoint can either be accessed by the requested user itself or by staff (admin) users.
A GET request returns full details of a specific user using the UserSerializer (like UserList).
A PUT request allows authorized users to edit all data of a specific user account. In practice, this will probably be restricted partially for normal users for individual projects.