Add environment variables to settings file
Add the default admin username to the settings file.
Add the default admin password to the settings file.
This variables will be used in the future to create a default superuser account when starting the project.

Also, the VERSION variable is added to the settings file instead of an extra file.