Create default admin user on startup
Create a default administrative user (superuser) on program startup.
This is done by calling a function as soon as the root URLs.py file is loaded (When this file is loaded, all apps are ready).
The create_admin_user() function then either creates a new superuser using the username and password set in the settings file, or it updates the existing user's password to the password defined in the settings file.

The imports in the user model file are changed and the is_staff variable is removed from the UserManager's create_user method, as it was not used.