Add default superuser creation error handling
Add error handling for the creation and updating process of the default superuser.

The unique attribute check was missing on changing an existing superuser's data. For this, this check now exists as the method unique_fail(), returning True if any user which is not the default superuser has the attributes given at settings.ADMIN_USER and .ADMIN_MAIL.