Add environment variables for SECRET_KEY and superuser data
Add the option to use an .env file for environment variables.
This requires the python-dotenv module (version 0.15.0).

The environment variables are loaded at the beginning of the project's settings file.
The .env file must be placed at the project's root folder.

Currently, the settings.SECRET_KEY as well as ADMIN_USER, ADMIN_MAIL, and ADMIN_PASSWORD are stored at the .env file.

An .env.example file is also added to the project. The .env file is added to .gitignore.